requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
python-dotenv
resend
aiosqlite
